// src/app/model/employee.type.ts
export type Employee = {
    id: string;
    name: string;
    surname: string;
    idNumber: string;
    hourlyRate: number;
    designation: string;
    leave: number;
    sickLeave: number;
    otherLeave: number;
    salary: number;
    taxRefNum: string;
    workStatus: string;
    statusChangeDate: string;
    storeId: string;
    storeDesignation: string;
    bankName: string;
    accountNumber: string;
    isEditing: boolean;
}